articles/examples/runtime/requirements.txt
# AgentCore Runtime Example
# https://docs.aws.amazon.com/bedrock-agentcore/

# Core AgentCore SDK
bedrock-agentcore>=1.0.0

# Agent framework (Strands is the simplest to start with)
strands-agents>=0.1.0

# AWS SDK for lower-level access
boto3>=1.34.0

# Environment management
python-dotenv>=1.0.0
